requirements.txt
# Core FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Google Cloud Services
google-cloud-sdk
google-cloud-aiplatform>=1.104.0
google-cloud-speech>=2.33.0
google-cloud-texttospeech>=2.27.0
google-cloud-firestore>=2.21.0
google-cloud-storage>=2.19.0

# Firebase
firebase-admin>=6.5.0

# Data validation and parsing
pydantic[email]>=2.5.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP client
httpx>=0.25.0

# Async support
anyio>=4.0.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Security
cryptography>=41.0.0

# Additional utilities
pathlib2>=2.3.7
typing-extensions>=4.8.0

# Audio processing (if needed)
librosa>=0.10.0
soundfile>=0.12.0

# Database (if using SQL databases)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Caching (if using Redis)
redis>=5.0.0

# Monitoring and logging
structlog>=23.0.0